change expected types for requests and responses to float64

+++ internal/base/base.go
@@ -6,19 +6,19 @@ func NewSimpleMathService() *SimpleMathService {
 	return &SimpleMathService{}
 }
 
-func (s *SimpleMathService) Add(a, b float32) float32 {
+func (s *SimpleMathService) Add(a, b float64) float64 {
 	return a + b
 }
 
-func (s *SimpleMathService) Sub(a, b float32) float32 {
+func (s *SimpleMathService) Sub(a, b float64) float64 {
 	return a - b
 }
 
-func (s *SimpleMathService) Mul(a, b float32) float32 {
+func (s *SimpleMathService) Mul(a, b float64) float64 {
 	return a * b
 }
 
-func (s *SimpleMathService) Div(a, b float32) float32 {
+func (s *SimpleMathService) Div(a, b float64) float64 {
 	if b == 0 {
 		return 0
 	}
